requirements.txt
kafka-python
faker
pyspark==4.0.1
psycopg2-binary